requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.11.2
webdriver-manager==4.0.0
dnspython==2.4.2
python-nmap==0.7.1
tldextract==3.4.4